refactor(viewer): split AuditRunsTab into components/admin/auditRuns
1,518-line tab becomes a folder: thin AuditRunsTab, gridMetrics.ts
(metric keys, cell values/labels/colours, quality flags), RunGrid,
TriggerForm, RunList, RunActions (cancel/validate/re-run/delete +
issue-bot badge), CellModals (cell history table on DataTable) and a
useAuditRuns hook owning the list/detail loads and the 5 s poll.
Markup, class names and strings unchanged.
