+++ src/frontend/src/components/admin/AuditTab.tsx
@@ -6,7 +6,7 @@ import type {AuditSubTab} from "./adminTabs";
 import AuditFilterBar from "./AuditFilterBar";
 import AuditLogTab from "./auditLog/AuditLogTab";
 import AuditOverviewTab from "./AuditOverviewTab";
-import AuditRunsTab from "./AuditRunsTab";
+import AuditRunsTab from "./auditRuns/AuditRunsTab";
 import CorpusTab from "./CorpusTab";
 import SchedulesTab from "./SchedulesTab";
